[model-gateway] adds default implementations to RouterTrait in mod.rs
